feat:网关路由前进行安全校验 #4164
