+++ src/gateway/core/lua/util/security_util.lua
@@ -15,34 +15,8 @@
 -- NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
 -- WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 -- SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-string = require("string")
-math = require("math")
-json = require("cjson.safe")
-uuid = require("resty.jit-uuid")
-resolver = require("resty.dns.resolver")
-ck = require("resty.cookie")
-http = require("resty.http")
-jwt = require("resty.jwt")
-stringUtil = require("util.string_util")
-ipUtil = require("util.ip_util")
-consulUtil = require("util.consul_util")
-logUtil = require("util.log_util")
-redisUtil = require("util.redis_util")
-oauthUtil = require("util.oauth_util")
-md5 = require("resty.md5")
-arrayUtil = require("util.array_util")
-cookieUtil = require("util.cookie_util")
-grayUtil = require("util.gray_util")
-urlUtil = require("util.url_util")
-tagUtil = require("util.tag_util")
-loadBalanceUtil = require("util.loadbalance_util")
-accessControlUtil = require("util.access_control_util")
-securityUtil = require("util.security_util")
-
-math.randomseed(ngx.now() * 1000)
-uuid.seed()
-
-local ok_table = {status = 0, data = true}
-
-response_ok = json.encode(ok_table)
-
+local _M = {}
+function _M:isSafe()
+    return true
+end
+return _M
