+++ src/gateway/core/map.conf
@@ -10,9 +10,6 @@
     default '';
   }
   
-  map $host $whitelist_deny {
-    default '';
-  }
   map $host $target {
     default '';
   }
